pkg/service/db_service/fscache_store_test.go
/*
Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserve.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

package db_service

import (
	"testing"

	"github.com/stretchr/testify/assert"

	"github.com/PaddlePaddle/PaddleFlow/pkg/common/database"
)

func TestGetFSCacheStore(t *testing.T) {
	database.InitMockDB()
	dbStore := GetFSCacheStore()
	fsCache, err := dbStore.Get("", "")
	assert.NotNil(t, err)
	assert.Nil(t, fsCache)
}
